refactor: move app_config into config module
